+++ spec/frise/loader_spec.rb
@@ -249,4 +249,12 @@ RSpec.describe Loader do
       }
     )
   end
+
+  it 'should validate schemas included deeper in the config file hierarchy' do
+    loader = Loader.new(exit_on_fail: true)
+    expect { loader.load(fixture_path('loader_test12.yml')) }.to output(
+      "1 config error(s) found:\n" \
+      " - At variable1.value2: missing required value\n"
+    ).to_stdout.and raise_error(SystemExit)
+  end
 end